Import minimum working version.
It uses libevdev to read keys from /dev/input/event* files that
represent keyboards. inotify is used to detect /dev/input/event*
files as they appear. Pre-existing files are opened on startup.
Each file is handled in a dedicated thread. All errors are handled
and logged cleanly.

Builds are based on cmake. The source code builds with gcc and clang
and is clean according to iwyu, clang-tidy and clang-analzyer.

There are no statistics yet. For now, it just prints debug messages
containing the detected keys.